test: Consolidate synthetic Examples EDM tests into test_event_data.py
test_measurement_creation/test_measurement_map_creation were pure synthetic
tests (no detector, no simulation) sitting inside test_truth_tracking.py,
which is otherwise about simulation-driven truth tracking. Move them into a
new Python/Examples/tests/test_event_data.py, mirroring the existing
Python/Core/tests/test_event_data.py for Core-module bindings, and fold the
track/track-container tests from test_track_container.py into the same file
rather than keeping a separate file for the same kind of coverage.
