+++ Python/Examples/tests/test_truth_tracking.py
@@ -624,140 +624,3 @@ def test_measurement_access(tmp_path, generic_detector_config):
 
         with failure_threshold(acts.logging.ERROR):
             seq.run()
-
-
-def test_measurement_creation():
-    meas_properties = [
-        {
-            "geometryId": acts.GeometryIdentifier(798),
-            "indices": [0],
-            "parameters": [1.0],
-            "covariance": [0.1],
-        },
-        {
-            "geometryId": acts.GeometryIdentifier(123),
-            "indices": [0, 1],
-            "parameters": [1.0, 2.0],
-            "covariance": [0.1, 0.1],
-        },
-        {
-            "geometryId": acts.GeometryIdentifier(456),
-            "indices": [0, 1, 4],
-            "parameters": [3.0, 4.0, 5.0],
-            "covariance": [0.2, 0.2, 0.2],
-        },
-    ]
-
-    container = acts.examples.MeasurementContainer()
-    container.reserve(3)
-    for meas_prop in meas_properties:
-        meas = container.emplaceMeasurement(**meas_prop)
-
-    for i in range(len(meas_properties)):
-        meas = container[i]
-        meas_prop = meas_properties[i]
-
-        dim = len(meas_prop["indices"])
-        assert meas.geometryId.value == meas_prop["geometryId"].value
-        assert [meas_prop["indices"][i] == meas.subspaceIndices[i] for i in range(dim)]
-        indices = meas_prop["indices"]
-        assert [
-            meas_prop["parameters"][i] == meas.fullParameters[indices[i]]
-            for i in range(dim)
-        ]
-        assert [
-            meas_prop["covariance"][i] == meas.fullCovariance[indices[i], indices[i]]
-            for i in range(dim)
-        ]
-
-    assert len(container) == 3
-
-    # Build a subset from indices 0 and 2 and verify it mirrors the container data
-    subset = acts.examples.MeasurementSubset(container, [0, 2])
-    assert len(subset) == 2
-
-    # Iteration covers exactly the selected measurements in order
-    subset_list = list(subset)
-    assert len(subset_list) == 2
-    assert subset_list[0].index == 0
-    assert subset_list[1].index == 2
-
-    # __getitem__ by subset position
-    assert subset[0].index == 0
-    assert subset[1].index == 2
-
-    # getMeasurement by original-container index
-    assert (
-        subset.getMeasurement(0).geometryId.value
-        == meas_properties[0]["geometryId"].value
-    )
-    assert (
-        subset.getMeasurement(2).geometryId.value
-        == meas_properties[2]["geometryId"].value
-    )
-
-    # Measurement data is consistent with the container entries
-    for pos, orig_idx in enumerate([0, 2]):
-        meas = subset[pos]
-        meas_prop = meas_properties[orig_idx]
-        dim = len(meas_prop["indices"])
-        assert meas.geometryId.value == meas_prop["geometryId"].value
-        indices = meas_prop["indices"]
-        assert [meas.subspaceIndices[i] == meas_prop["indices"][i] for i in range(dim)]
-        assert [
-            meas.fullParameters[indices[i]] == meas_prop["parameters"][i]
-            for i in range(dim)
-        ]
-
-
-def test_measurement_map_creation():
-    from acts.examples import (
-        MeasurementParticlesMap,
-        MeasurementSimHitsMap,
-        ParticleMeasurementsMap,
-        SimBarcode,
-        SimHitMeasurementsMap,
-    )
-
-    # MeasurementSimHitsMap: meas 0 → simhits {10, 11}, meas 1 → simhit {20}
-    m = MeasurementSimHitsMap()
-    m.insert(0, 10)
-    m.insert(0, 11)  # same key — multi-map
-    m.insert(1, 20)
-    assert len(m) == 3
-
-    assert 0 in m
-    assert 2 not in m
-
-    vals = m.valuesFor(0)
-    assert sorted(vals) == [10, 11]
-    assert m.valuesFor(1) == [20]
-    assert m.valuesFor(99) == []
-
-    pairs = list(m)
-    assert len(pairs) == 3
-    assert all(isinstance(k, int) and isinstance(v, int) for k, v in pairs)
-
-    inv = m.invert()
-    assert isinstance(inv, SimHitMeasurementsMap)
-    assert len(inv) == 3
-    assert inv.valuesFor(10) == [0]
-    assert inv.valuesFor(11) == [0]
-    assert inv.valuesFor(20) == [1]
-
-    # MeasurementParticlesMap: meas 0 came from two particles, meas 1 from one
-    bc0 = SimBarcode()
-    bc0.particle = 1
-    bc1 = SimBarcode()
-    bc1.particle = 2
-    mp = MeasurementParticlesMap()
-    mp.insert(0, bc0)
-    mp.insert(0, bc1)  # same measurement, two particles
-    mp.insert(1, bc0)
-    assert len(mp) == 3
-
-    assert mp.valuesFor(0) == [bc0, bc1]
-
-    inv_p = mp.invert()
-    assert isinstance(inv_p, ParticleMeasurementsMap)
-    assert len(inv_p) == 3
